% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_1.R
\name{is_permutation}
\alias{is_permutation}
\title{Check if a string is a permutation of the other, vectorized}
\usage{
is_permutation(str1, str2)
}
\arguments{
\item{str1}{The first string to check against.}

\item{str2}{The second string to check against.}
}
\description{
Check if a string is a permutation of the other, vectorized
}
