% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_1.R
\name{palindrome_permutation}
\alias{palindrome_permutation}
\title{Check that a string is a permutation in linear time, vectorized}
\usage{
palindrome_permutation(strs)
}
\arguments{
\item{strs}{The potential permuted palindromes.}
}
\value{
\code{TRUE} for permuted palindromes, \code{FALSE} otherwise.
}
\description{
Check that a string is a permutation in linear time, vectorized
}
