% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_1.R
\name{unique_characters}
\alias{unique_characters}
\title{Check if a string has all unique characters, vectorized}
\usage{
unique_characters(strs)
}
\arguments{
\item{strs}{The string to check}
}
\description{
Check if a string has all unique characters, vectorized
}
