% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_1.R
\name{unique_strings}
\alias{unique_strings}
\title{Check if a vector of strings has distinct entries}
\usage{
unique_strings(strs)
}
\arguments{
\item{strs}{The strings to check}
}
\description{
A variation of unique_characters as it was in the book
}
