% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_1.R
\name{URLify}
\alias{URLify}
\title{Replace every space with '%20', vectorized}
\usage{
URLify(str)
}
\arguments{
\item{str}{The string to modify.}
}
\value{
The URLified string.
}
\description{
This solution is not in spirit of the exercise, R doesn't give this level of
control and I don't believe it should.
}
