% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_1.R
\name{unique_strings}
\alias{unique_strings}
\title{Check if a vector of strings has distinct entries}
\usage{
unique_strings(strs)
}
\arguments{
\item{strs}{The strings to check}
}
\description{
A variation of \code{unique_characters()} as it is in the book using a vector
of strings to be checked instead of characters and using a dictionary instead
of \code{unique()}.
}
